complete statistic api on statistic_view.py file

+++ my_app/models.py
@@ -44,8 +44,8 @@ class Statistic(models.Model):
     user_id = models.ForeignKey(User, on_delete=models.CASCADE)
     question_id = models.OneToOneField(Question, on_delete=models.CASCADE, related_name='statistics')
     category = models.CharField(max_length=50,choices=CategoryChoices.choices, default=CategoryChoices.EDEMIREM)
-    correct_answers = models.PositiveIntegerField()
-    incorrect_answers = models.PositiveIntegerField()
+    correct_answers = models.FloatField()
+    incorrect_answers = models.FloatField()
 
     def __str__(self):
         return self.user_id.username
